взрыв доступен в базовом классе, опционален, и разделён на методы
