+++ Source/MyProject/PhysicsProjectile.cpp
@@ -8,7 +8,6 @@
 #include "CollisionShape.h"
 #include "CollisionQueryParams.h"
 #include "Engine/EngineTypes.h"
-#include "Components/AudioComponent.h"
 
 
 APhysicsProjectile::APhysicsProjectile()
@@ -18,16 +17,6 @@ APhysicsProjectile::APhysicsProjectile()
 	TrailEffect = CreateDefaultSubobject<UParticleSystemComponent>(TEXT("Trail effect"));
 	TrailEffect->SetupAttachment(RootComponent);
 	TrailEffect->SetAutoActivate(false);
-
-	// инициализируем визуальный эффект взрыва
-	ExplodeEffect = CreateDefaultSubobject<UParticleSystemComponent>(TEXT("Explode effect"));
-	ExplodeEffect->SetupAttachment(RootComponent);
-	ExplodeEffect->SetAutoActivate(false);
-
-	// инициализируем аудио эффект взрыва
-	ExplodeAudio = CreateDefaultSubobject<UAudioComponent>(TEXT("Explode Sound"));
-	ExplodeAudio->SetupAttachment(RootComponent);
-	ExplodeAudio->SetAutoActivate(false);
 }
 
 void APhysicsProjectile::Start()
@@ -65,7 +54,10 @@ void APhysicsProjectile::Move()
 
 		if (TragectoryPointIndex >= CurrentTrajectory.Num())
 		{
-			Explode();
+			// проверяем доступность взрыва в классе
+			if (ExplodeAvailable)
+				Explode();
+			
 			Destroy();
 		}
 		else
@@ -74,85 +66,4 @@ void APhysicsProjectile::Move()
 			SetActorRotation(newRotation);
 		}
 	}
-}
-
-void APhysicsProjectile::Explode()
-{
-	// воспроизводим звук
-	if (ExplodeAudio)
-		ExplodeAudio->Play();
-
-	// запускаем визуальный эффект
-	if (ExplodeEffect)
-		ExplodeEffect->ActivateSystem();
-
-	FVector startPos = GetActorLocation();
-	FVector endPos = startPos + FVector(0.1f);
-	FCollisionShape Shape = FCollisionShape::MakeSphere(ExplodeRadius);
-
-	FCollisionQueryParams params = FCollisionQueryParams::DefaultQueryParam;
-	params.AddIgnoredActor(this);
-	params.bTraceComplex = true;
-	params.TraceTag = "Explode Trace";
-
-	TArray<FHitResult> AttackHit;
-	FQuat Rotation = FQuat::Identity;
-
-	bool sweepResult = GetWorld()->SweepMultiByChannel
-	(
-		AttackHit,
-		startPos,
-		endPos,
-		Rotation,
-		ECollisionChannel::ECC_Visibility,
-		Shape,
-		params
-	);
-
-	//GetWorld()->DebugDrawTraceTag = "Explode Trace";
-
-	if (sweepResult)
-	{
-		for (FHitResult hitResult : AttackHit)
-		{
-			AActor* otherActor = hitResult.GetActor();
-
-			if (!otherActor)
-				continue;
-
-			IDamageTaker* damageTakerActor = Cast<IDamageTaker>(otherActor);
-
-			if (damageTakerActor)
-			{
-				FDamageData damageData;
-				damageData.DamageValue = Damage;
-				damageData.Instigator = GetOwner();
-				damageData.DamageMaker = this;
-				damageTakerActor->TakeDamage(damageData);
-			}
-			else
-			{
-				UPrimitiveComponent* mesh =	Cast<UPrimitiveComponent>(otherActor->GetRootComponent());
-
-				if (mesh)
-				{
-					if (mesh->IsSimulatingPhysics())
-					{
-						FVector forceVector = otherActor->GetActorLocation() - GetActorLocation();
-						forceVector.Normalize();
-
-						mesh->AddImpulse(forceVector * PushForce, NAME_None, true);
-					}
-				}
-			}
-		}
-	}
-}
-
-void APhysicsProjectile::OnMeshOverlapBegin(class UPrimitiveComponent* OverlappedComp, class AActor* OtherActor, class UPrimitiveComponent* OtherComp, int32 OtherBodyIndex, bool bFromSweep, const FHitResult& SweepResult)
-{
-	// добавляем взрыв при взаимодействии с объектами
-	Explode();
-
-	Super::OnMeshOverlapBegin(OverlappedComp, OtherActor, OtherComp, OtherBodyIndex, bFromSweep, SweepResult);
 }
