+++ Source/MyProject/Projectile.cpp
@@ -5,15 +5,31 @@
 #include "DamageTaker.h"
 #include "GameStruct.h"
 #include "Components/PrimitiveComponent.h"
+#include "Components/AudioComponent.h"
+#include "Particles/ParticleSystemComponent.h"
+
+
 
 AProjectile::AProjectile()
 {
 	PrimaryActorTick.bCanEverTick = false;
+
 	USceneComponent* sceeneCpm = CreateDefaultSubobject<USceneComponent>(TEXT("Root"));
 	RootComponent = sceeneCpm;
+
 	Mesh = CreateDefaultSubobject<UStaticMeshComponent>(TEXT("Mesh"));
 	Mesh->SetupAttachment(RootComponent);
 	Mesh->OnComponentBeginOverlap.AddDynamic(this, &AProjectile::OnMeshOverlapBegin);
+
+	// инициализируем визуальный эффект взрыва
+	ExplodeEffect = CreateDefaultSubobject<UParticleSystemComponent>(TEXT("Explode effect"));
+	ExplodeEffect->SetupAttachment(RootComponent);
+	ExplodeEffect->SetAutoActivate(false);
+
+	// инициализируем аудио эффект взрыва
+	ExplodeAudio = CreateDefaultSubobject<UAudioComponent>(TEXT("Explode Sound"));
+	ExplodeAudio->SetupAttachment(RootComponent);
+	ExplodeAudio->SetAutoActivate(false);
 }
 
 void AProjectile::Start()
@@ -28,30 +44,15 @@ void AProjectile::OnMeshOverlapBegin(class UPrimitiveComponent* OverlappedComp, 
 
 	if (OtherActor != owner && OtherActor != ownerByOwner)
 	{
-		IDamageTaker* damageTakerActor = Cast<IDamageTaker>(OtherActor);
-		if (damageTakerActor)
-		{
-			FDamageData damageData;
-			damageData.DamageValue = Damage;
-			damageData.Instigator = ownerByOwner;
-			damageData.DamageMaker = this;
-			damageTakerActor->TakeDamage(damageData);
-		}
-		else
-		{
-			UPrimitiveComponent* mesh =	Cast<UPrimitiveComponent>(OtherActor->GetRootComponent());
-			if (mesh)
-			{
-				if (mesh->IsSimulatingPhysics())
-				{
-					FVector forceVector = OtherActor->GetActorLocation() - GetActorLocation();
-					forceVector.Normalize();
-					//mesh->AddImpulse(forceVector * PushForce, NAME_None, true);
-					mesh->AddForce(forceVector * PushForce, NAME_None, true);
-				}
-			}
-		}
+		if (!MakeDamageTo(OtherActor))
+			PushActor(OtherActor);
 
+		// проверяем доступность взрыва в классе
+		if (ExplodeAvailable)
+		{
+			// взрыыв при любом столкновении
+			Explode();
+		}
 
 		Destroy();
 	}
@@ -62,5 +63,92 @@ void AProjectile::Move()
 {
 	FVector nextPosition = GetActorLocation() + GetActorForwardVector() * MoveSpeed	* MoveRate;
 	SetActorLocation(nextPosition);
+}
+
+void AProjectile::Explode()
+{
+	// воспроизводим звук
+	if (ExplodeAudio)
+		ExplodeAudio->Play();
+
+	// запускаем визуальный эффект
+	if (ExplodeEffect)
+		ExplodeEffect->ActivateSystem();
+
+	FVector startPos = GetActorLocation();
+	FVector endPos = startPos + FVector(0.1f);
+	FCollisionShape Shape = FCollisionShape::MakeSphere(ExplodeRadius);
+
+	FCollisionQueryParams params = FCollisionQueryParams::DefaultQueryParam;
+	params.AddIgnoredActor(this);
+	params.bTraceComplex = true;
+	params.TraceTag = "Explode Trace";
+
+	TArray<FHitResult> AttackHit;
+	FQuat Rotation = FQuat::Identity;
+
+	bool sweepResult = GetWorld()->SweepMultiByChannel
+	(
+		AttackHit,
+		startPos,
+		endPos,
+		Rotation,
+		ECollisionChannel::ECC_Visibility,
+		Shape,
+		params
+	);
+
+	GetWorld()->DebugDrawTraceTag = "Explode Trace";
+
+	if (sweepResult)
+	{
+		for (FHitResult hitResult : AttackHit)
+		{
+			AActor* otherActor = hitResult.GetActor();
+
+			if (!otherActor)
+				continue;
+
+			if(!MakeDamageTo(otherActor))
+				PushActor(otherActor);
+		}
+	}
+}
+
+bool AProjectile::MakeDamageTo(AActor* otherActor)
+{
+	IDamageTaker* damageTakerActor = Cast<IDamageTaker>(otherActor);
+
+	if (!damageTakerActor)
+		return false;
+
+	FDamageData damageData;
+	damageData.DamageValue = Damage;
+	damageData.Instigator = GetOwner();
+	damageData.DamageMaker = this;
+	damageTakerActor->TakeDamage(damageData);
+
+	return true;
+}
+
+bool AProjectile::PushActor(AActor* otherActor)
+{
+	if (!otherActor)
+		return false;
+
+	UPrimitiveComponent* mesh = Cast<UPrimitiveComponent>(otherActor->GetRootComponent());
+
+	if (!mesh)
+		return false;
+
+	if (!mesh->IsSimulatingPhysics())
+		return false;
+
+	FVector forceVector = otherActor->GetActorLocation() - GetActorLocation();
+	forceVector.Normalize();
+
+	mesh->AddImpulse(forceVector * PushForce, NAME_None, true);
+
+	return true;
 }
